finetuning/requirements-mac.txt
torch
torchvision
torchaudio
transformers
datasets
accelerate
peft
numpy
pandas
tqdm
pillow
matplotlib
jupyterlab
ipywidgets
sentencepiece
trl
qwen-vl-utils
sacrebleu
rouge-score
codebleu
scikit-image
playwright
